+++ neo/idlib/math/Simd_Generic.cpp
@@ -122,7 +122,7 @@ void VPCALL idSIMD_Generic::MinMax( idVec3 &min, idVec3 &max, const idDrawVert *
 idSIMD_Generic::Memcpy
 ================
 */
-void VPCALL idSIMD_Generic::Memcpy( void *dst, const void *src, const int count ) {
+void VPCALL idSIMD_Generic::Memcpy( void *dst, const void *src, const size_t count ) {
 	memcpy( dst, src, count );
 }
 
@@ -131,7 +131,7 @@ void VPCALL idSIMD_Generic::Memcpy( void *dst, const void *src, const int count 
 idSIMD_Generic::Memset
 ================
 */
-void VPCALL idSIMD_Generic::Memset( void *dst, const int val, const int count ) {
+void VPCALL idSIMD_Generic::Memset( void *dst, const int val, const size_t count ) {
 	memset( dst, val, count );
 }
