+++ neo/idlib/math/Simd_Generic.h
@@ -47,8 +47,8 @@ public:
 	virtual	void VPCALL MinMax( idVec3 &min,		idVec3 &max,			const idDrawVert *src,	const int count );
 	virtual	void VPCALL MinMax( idVec3 &min,		idVec3 &max,			const idDrawVert *src,	const triIndex_t *indexes,		const int count );
 
-	virtual void VPCALL Memcpy( void *dst,			const void *src,		const int count );
-	virtual void VPCALL Memset( void *dst,			const int val,			const int count );
+	virtual void VPCALL Memcpy( void *dst,			const void *src,		const size_t count );
+	virtual void VPCALL Memset( void *dst,			const int val,			const size_t count );
 
 	virtual void VPCALL BlendJoints( idJointQuat *joints, const idJointQuat *blendJoints, const float lerp, const int *index, const int numJoints );
 	virtual void VPCALL BlendJointsFast( idJointQuat *joints, const idJointQuat *blendJoints, const float lerp, const int *index, const int numJoints );
